Initial implementation of i18n with jslingui'
Refs #31

jslingui is pinned to the 2.0.1 release due to later releases depending
on a alpha version of Babel.

Compiled messages are in their [lang]/messages.js files. It remains to
be seen whether we want these committed to repo or not. I have included
them for now.

Adds a couple of new build commands:
 - npm run jslingui -- extract // Extracts all strings from codebase.
 - npm run jslingui -- compile --strict // Compiles translations into
 optimized, importable code.
 - npm run jslingui -- add-locale [en] // Adds a new locale to create
 translation messages for.